рефакторинг и чистка
