+++ project/src/components/pages/favorites/favorites-locations/favorites-locations.jsx
@@ -5,7 +5,7 @@ import {Link} from 'react-router-dom';
 import FavoriteOffer from '../favorite-offer/favorite-offer';
 
 
-function FavoritesLocations({city, offers, listUpdateHandler}) {
+function FavoritesLocations({city, offers, updateOffers}) {
   return (
     <li className="favorites__locations-items">
       <div className="favorites__locations locations locations--current">
@@ -28,7 +28,7 @@ function FavoritesLocations({city, offers, listUpdateHandler}) {
               previewImage={offer.previewImage}
               isFavorite={offer.isFavorite}
               isPremium={offer.isPremium}
-              listUpdateHandler={listUpdateHandler}
+              favButtonClickHandler={updateOffers}
             />
           ))
         }
@@ -42,7 +42,7 @@ FavoritesLocations.propTypes = {
   offers: PropTypes.arrayOf(
     PropTypes.shape(offerFullProp),
   ),
-  listUpdateHandler: PropTypes.func.isRequired,
+  updateOffers: PropTypes.func.isRequired,
 };
