+++ project/src/components/pages/main/city-offer/city-offer.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import {useDispatch} from 'react-redux';
+// import {useDispatch} from 'react-redux';
 import {offerBasicProp} from '../../../ui/offer/offer-prop';
 import Offer from '../../../ui/offer/offer';
-import {updateFavoriteStatus} from '../../../../store/api-action';
+// import {updateFavoriteStatus} from '../../../../store/api-action';
 
 
 const SpecialClassName = {
@@ -13,21 +13,21 @@ const SpecialClassName = {
 
 
 function CityOffer(props) {
-  const {id, isFavorite} = props;
+  // const {id, isFavorite} = props;
 
-  const dispatch = useDispatch();
+  // const dispatch = useDispatch();
 
-  const favButtonClickHandler = (evt) => {
-    evt.preventDefault();
+  // const favButtonClickHandler = (evt) => {
+  //   evt.preventDefault();
 
-    dispatch(updateFavoriteStatus(id, Number(!isFavorite)));
-  };
+  //   dispatch(updateFavoriteStatus(id, Number(!isFavorite)));
+  // };
 
   return (
     <Offer
       cardClassName={SpecialClassName.CARD_CLASS}
       imageWrapperClassName={SpecialClassName.IMAGE_WRAPPER_CLASS}
-      favButtonClickHandler={favButtonClickHandler}
+      // favButtonClickHandler={favButtonClickHandler}
       {...props}
     />
   );
