+++ project/src/components/pages/main/offers-list/offers-list.jsx
@@ -29,7 +29,6 @@ function OffersList() {
   const currentSortType = useSelector(getCurrentSortType);
 
   const sortedOffers = sortOffers(currentCityoffers, currentSortType);
-  // console.log(sortedOffers);
   const dispatch = useDispatch();
 
   return (
